lacework_agent_new/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
flask>=2.0.0
# Optional: uncomment to enable OpenAI Agents SDK support
# openai-agents
